Amended SQL for Step 5-5 due to data bug
Whilst creating a visualisation for this dataset, I discovered some Shipment Dates had values incorrectly before their respective Order Dates, where Order Date is close to end of year and Shipment Date is close to start of the same year.

As the Shipment Date value is allocated randomly anyway, I added an update SQL at the end of 5-5 to identify those records and +1 year to Shipment Date.

Updated the Excel files as well for the new output.